This was causing a big headache and I think for things like this are what people discourage the use of structs... it is always interesting taking a look at why

+++ common/src/SpriteManager.c
@@ -26,7 +26,7 @@ void SpriteManagerReset() {
 	//place all sprites on the pool
 	sprite_manager_sprites_pool[0] = N_SPRITE_MANAGER_SPRITES;
 	for(i = 0; i != N_SPRITE_MANAGER_SPRITES; ++i) {
-		sprite_manager_sprites[i] = &sprite_manager_sprites_mem[sizeof(struct Sprite) * i];
+		sprite_manager_sprites[i] = &sprite_manager_sprites_mem[sizeof(struct Sprite) * (UINT16)i];
 
 		sprite_manager_sprites_pool[i + 1] = i;
